Add files via upload
Initial code for the Mancala game board.

Board.java features two classes, Board and Hollow. A standard board is made up of 12 middle Hollows, and two end Hollows.

Functionality of board so far is mostly constructors to set up the hollows, and could allow user to initialize with a certain number of hollows. Constructors are basic so far, just to give us an idea of some things we can do with the properties of the board.

Class Hollow features methods that related the number of game pieces it is holding. Has a couple set and get functions so far.